requirements.txt
streamlit== 1.36.0
pandas==2.2.2
matplotlib== 3.9.1
mysql-connector-python
streamlit
seaborn


